+++ src/main/java/com/badals/admin/domain/projection/ShipmentItemSummary.java
@@ -2,10 +2,12 @@ package com.badals.admin.domain.projection;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 public interface ShipmentItemSummary {
     Integer getId();
     String getTrackingNum();
-    Long getTotal();
+    BigDecimal getTotal();
     String getStatus();
     Long getProcessed();
     String getReference();
